requirements.txt
pycld2>=0.31
requests
pandas>=0.24.2
opencv-python>=4.1.0
Pillow>=6.2.1
CMake
dlib>=19.17.0
face_recognition>=1.2.3
torch>=1.0.0
numpy>=1.13
tqdm
torchvision>=0.2.2
pathlib
m3inference
keras
tensorflow
imutils
gender_guesser